datacollection/requirements.txt
azure-storage-queue==12.1.1
requests==2.23.0